no building required
I got rid of the doc folder, because the docs were basicly all blank, the test folder, because I'm not using them and I dout anyone else ever will, and the makefiles, because this really doesn't need to be an executable in my opinion.  my goal is to greatly simplify the directory structure and installation process as I merge Alex's code with mine into a streamlined pipeline.  is that redundant?  a streamed pipeline.  a streamlined pipe.