for OMEGA users
I made several changes to account for the differences between OMEGA shots and NIF shots, fixing problems that arose when I did the refactoring.  specficly:
I don't require OMEGA shots to be found in the shot table
I allow the user to supply a shell material when it isn't available in the shot table
I print a useful error message when the shell material isn't known and the user didn't supply it.
I make sure that WRFs without a specific position show up with no position, rather than position "None".