standardize analysis formats
I went thru and reprocessd all the WRFs for which the analysis was already done.  it helpd me find some bugs in my code, and more importantly generated report.xlsx files for all wedges ever.  I didn't account for any doping in the ablators because I didn't want to, so that changed the inferd ρRs slitely, but I think it's probably all within error bars.