use more faithful report template
I was messing with the templates because I thaut there was some magical formatting thing that was causing the Archive to reject my reports.  there wasn't as it turns out; the format saved by openpyxl is just *slitely* different from the format saved by Excel; they're both readable by openpyxl and they're both readable by Excel, but only the Excel version is readable by the Archive.  and the Archive needs to read the spreadsheet file to infer the shot and DIM.  well, anyway, it turns out you can still use these new autogenerated report spreadsheets, but you have to open and save them in Excel first.  which I think is an acceptable workaround.