take user-specified ne and Te
this required me to revise a lot of the stopping tables.  I also adjusted how it does the error bars, so most of the existing tables are now useless.

Oh, I also found out that these tables titled "stopping range" are not stopping range, but the correspondance between rhoR and E.  meaning these calculations were rong.  so – completely intentionally – I changed the tables to be what they actually claim to be.

I'm realizing that this is not very sustainable anyway, so I really aut to just figure out how to use Alex's code.