forgive spreadsheets with partial shot names
previusly when a traveler spreadsheet only had part of a shot name, it would stick "S01a" on the end, and then if the actual shot was not the first in the series, it would get mad that the shot names don't match. now it uses "SXXx" to indicate an indeterminate shot name, which matches correctly with the full shot name.

this is the real commit.  I have no idea how that random aux_info.csv got deleted in the git tree.