make it better at determining shot number
now it queries the shot_info table using the same shot number it pulls from the folder/filename and labels the plots with, rather than trying to parse the folder *as* a shot number (which was way unnecessarily strict and put a lot of silly constraints on the folder name).  the error messages are also a bit more informative in general now.