report files for individual spectra
I've been putting this off too long.  now when you run the code it will generate report spreadsheets containing all the spectral information for each WRF, in addition to the ones summarizing the key scalars on each DIM, that can be uploaded to the NIF archive.  remains to be seen if the archive will actually take them.